refactor(icons): split icon glyphs into a constants-only file
react-refresh/only-export-components warned on icons.tsx (17 warnings)
because it mixed FC components with plain JSX-element constants in one
file. Moved the constants to iconGlyphs.tsx; icons.tsx now exports only
components.
