+++ src/components/ReorderControls/ReorderControls.tsx
@@ -1,5 +1,5 @@
 import Button from '@components/Button'
-import { iconChevronDown, iconChevronUp, iconRemove } from '@components/icons'
+import { iconChevronDown, iconChevronUp, iconRemove } from '@components/iconGlyphs'
 import type { FC } from 'react'
 
 export interface ReorderControlsProps {
