+++ src/pages/admin/RecipeList/RecipeList.tsx
@@ -7,8 +7,6 @@ import {
 } from '@api/recipes'
 import ConfirmDialog from '@components/ConfirmDialog'
 import {
-  IconPlus,
-  IconPreview,
   iconDelete,
   iconDocument,
   iconEdit,
@@ -16,7 +14,8 @@ import {
   iconRetry,
   iconUnpublish,
   iconWarning,
-} from '@components/icons'
+} from '@components/iconGlyphs'
+import { IconPlus, IconPreview } from '@components/icons'
 import Link from '@components/Link'
 import StateBox from '@components/StateBox'
 import StatusBadge from '@components/StatusBadge'
