+++ src/pages/admin/RecipePreview/RecipePreview.tsx
@@ -1,7 +1,8 @@
 import { isSessionError } from '@api/auth'
 import { fetchRecipeByIdAdmin, publishRecipe } from '@api/recipes'
 import Button from '@components/Button'
-import { IconPreview, iconEdit, iconNotFound, iconPublish, iconViewPublic } from '@components/icons'
+import { iconEdit, iconNotFound, iconPublish, iconViewPublic } from '@components/iconGlyphs'
+import { IconPreview } from '@components/icons'
 import Link from '@components/Link'
 import Loading from '@components/Loading'
 import RecipeDetailView from '@components/RecipeDetailView'
